UIOR-123 fix related invoices on POL
